fix: godot 3 does not support longs on method parameters - got the error 'Failed getting method ID saveGame' when running on android followed by crash

+++ app/src/main/java/com/jacobibanez/godot/gpgs/GodotGooglePlayGameServices.kt
@@ -134,9 +134,9 @@ class GodotGooglePlayGameServices(
         fileName: String,
         description: String,
         saveData: ByteArray,
-        playedTimeMillis: Long,
-        progressValue: Long
-    ) = snapshotsProxy.saveGame(fileName, description, saveData, playedTimeMillis, progressValue)
+        playedTimeMillis: Int,
+        progressValue: Int
+    ) = snapshotsProxy.saveGame(fileName, description, saveData, playedTimeMillis.toLong(), progressValue.toLong())
 
     @UsedByGodot
     fun loadGame(fileName: String) = snapshotsProxy.loadGame(fileName)
